P1: Corregido problema en el cálculo del MSE.

+++ p1/practica1.m
@@ -680,5 +680,5 @@ function error = mse(I, G)
 % I     Imagen original
 % G     Imagen observada
 
-E = ((I - G) .^ 2);
+E = double(I - G) .^ 2;
 error = sum(E(:)) / prod(size(I));
